third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/CustomElementRegistry-upgrade-expected.txt

This is a testharness.js-based test.
[FAIL] upgrade should not upgrade a candidate element not associated with a registry
  assert_equals: expected object "[object CustomElementRegistry]" but got null
Harness: the test ran to completion.
